+++ src/index.tsx
@@ -1,6 +1,5 @@
 import React, { FC, HTMLAttributes, useState, useCallback } from 'react';
 import styled from 'styled-components';
-import TextareaAutosize from 'react-autosize-textarea';
 import useDimensions from 'react-use-dimensions';
 import { ConvertCard } from './styled/ConvertCard';
 import { IconX } from './IconX';
@@ -21,8 +20,7 @@ import {
   IMaxContentLengthIndicator,
   MaxContentLengthIndicator,
 } from './MaxContentLengthIndicator';
-
-export { IInOption, IOutOption, InOptions, OutOptions };
+import { Textarea } from './styled/Textarea';
 
 const ConvertCardContent = styled.div`
   width: 100%;
@@ -39,36 +37,6 @@ const OptionsContainer = styled.div`
 
   @media (max-width: 576px) {
     display: none;
-  }
-`;
-
-type ExampleTextProps = {
-  smallerFont?: boolean;
-  showCopyCursor?: boolean;
-};
-
-const Textarea = styled(TextareaAutosize)<ExampleTextProps>`
-  box-sizing: border-box;
-  border: 0;
-  resize: none;
-  flex-grow: 1;
-  color: ${props => (props.theme.main === 'dark' ? '#fff' : '#14213d')};
-  background-color: transparent;
-  font-family: ${props =>
-    props.theme && props.theme.font ? props.theme.font : 'Roboto'};
-  font-size: ${props => (props.smallerFont ? '1.2em' : '1.61em')};
-  ${props => (props.showCopyCursor ? 'cursor: text;' : '')};
-  width: 100%;
-
-  ::placeholder {
-    color: ${props =>
-      props.theme.main === 'dark'
-        ? 'hsl(221, 51%, 64%)'
-        : 'rgba(20, 33, 61, 0.4)'};
-  }
-
-  :focus {
-    outline: none;
   }
 `;
